Code review fixes.
